requirements.txt
polars==1.0.0
pandas==2.2.0
scikit-learn==1.4.0
seaborn==0.13.0
folium==0.15.0
joblib==1.3.2
python-dotenv==1.0.1
streamlit==1.31.0
matplotlib==3.8.0